Added 16F1455 and fixed some typos
This seems buggy...

+++ usb/include/usb_cdc.h
@@ -227,15 +227,15 @@ struct cdc_serial_state_notification {
 	struct cdc_notification_header header;
 	union {
 		struct {
-			uint16_t bRxCarrier : 1; /**< Indicates DCD */
-			uint16_t bTxCarrier : 1; /**< Indicates DSR */
-			uint16_t bBreak : 1;
-			uint16_t bRingSignal : 1;
-			uint16_t bFraming : 1;
-			uint16_t bParity : 1;
-			uint16_t bOverrun : 1;
-			uint16_t : 1;
-			uint16_t : 8; /* XC8 can't handle a 9-bit bitfield */
+			uint8_t bRxCarrier : 1; /**< Indicates DCD */
+			uint8_t bTxCarrier : 1; /**< Indicates DSR */
+			uint8_t bBreak : 1;
+			uint8_t bRingSignal : 1;
+			uint8_t bFraming : 1;
+			uint8_t bParity : 1;
+			uint8_t bOverrun : 1;
+			uint8_t : 1;
+			uint8_t : 8; /* XC8 can't handle a 9-bit bitfield */
 		} bits;
 		uint16_t serial_state;
 	} data;
